Move common C# code to new C# base class
